DOC: Add short descriptions for all function.
