+++ src/file_input.cpp
@@ -83,11 +83,6 @@ std::string replace_spaces_with_commas(std::string instring) {
 }
 
 // -------------------------------------------------------------------
-//
-//
-// -------------------------------------------------------------------
-
-// -------------------------------------------------------------------
 // Read a string, clean it up, and report error if length is 0
 // -------------------------------------------------------------------
