+++ src/time.cpp
@@ -71,7 +71,7 @@ void Times::calc_dt() {
 }
 
 // -----------------------------------------------------------------------------
-//
+// Get the current time as a double
 // -----------------------------------------------------------------------------
 
 double Times::get_current() {
@@ -79,7 +79,7 @@ double Times::get_current() {
 }
 
 // -----------------------------------------------------------------------------
-//
+// Get the end time as a double
 // -----------------------------------------------------------------------------
 
 double Times::get_end() {
@@ -87,7 +87,7 @@ double Times::get_end() {
 }
 
 // -----------------------------------------------------------------------------
-//
+// Get the current time as a string
 // -----------------------------------------------------------------------------
 
 std::string Times::get_YMD_HMS() {
@@ -95,7 +95,7 @@ std::string Times::get_YMD_HMS() {
 }
 
 // -----------------------------------------------------------------------------
-//
+// Get the intermediate stopping time
 // -----------------------------------------------------------------------------
 
 double Times::get_intermediate() {
@@ -103,7 +103,7 @@ double Times::get_intermediate() {
 }
 
 // -----------------------------------------------------------------------------
-//
+// Get the current dt
 // -----------------------------------------------------------------------------
 
 float Times::get_dt() {
@@ -111,7 +111,7 @@ float Times::get_dt() {
 }
 
 // -----------------------------------------------------------------------------
-//
+// Get the orbit time, needed to update planetary characteristics
 // -----------------------------------------------------------------------------
 
 float Times::get_orbittime() {
@@ -119,7 +119,7 @@ float Times::get_orbittime() {
 }
 
 // -----------------------------------------------------------------------------
-//
+// Get the julian day
 // -----------------------------------------------------------------------------
 
 double Times::get_julian_day() {
@@ -127,7 +127,7 @@ double Times::get_julian_day() {
 }
 
 // -----------------------------------------------------------------------------
-//
+// Set the end time using a vector of year, month, day, hour, min, sec, msec
 // -----------------------------------------------------------------------------
 
 void Times::set_end_time(std::vector<int> itime) {
